refactor(schedule): move the interval duration vocabulary onto the shared protocol word table
The vocabulary was its own protocol module, which put it outside the thin CLI's
static import graph: check-cli-structure walks packages/cli/src/index.ts and only
admits the cross-package files in allowedStaticGraph, so the CLI's `--every`
parser reaching it reds G-CLI-STRUCTURE.

The three consumers have exactly one reachable layer in common. Kernel is not it:
the same walk names kernel domain modules and the kernel barrel as violations, and
the GUI renderer's no-restricted-imports bans every path with a kernel segment, so
kernel is unreachable from both ends at once. daemon-protocol-vocabulary.ts is the
protocol's shared word table, already reachable by all three, so the codec moves
there and the standalone module is deleted; no gate surface changes. The status-word
ratchet is untouched -- it judges *Words mirrors, which these are not.

Behaviour is unchanged: same units, same bounds, same five call sites.
